split out FormattingControls and TransactionsFilterControls
